Struct rust_sodium::crypto::secretbox::xsalsa20poly1305::Nonce

#[must_use]
pub struct Nonce(pub [u8; NONCEBYTES]);

Nonce for symmetric authenticated encryption

Methods

impl Nonce

fn from_slice(bs: &[u8]) -> Option<Nonce>

from_slice() creates an object from a byte slice

This function will fail and return None if the length of the byte-s;ice isn't equal to the length of the object

fn increment_le(&self) -> Nonce

increment_le() treats the nonce as an unsigned little-endian number and returns an incremented version of it.

WARNING: this method does not check for arithmetic overflow. It is the callers responsibility to ensure that any given nonce value is only used once. If the caller does not do that the cryptographic primitives in rust_sodium will not uphold any security guarantees (i.e. they will break)

fn increment_le_inplace(&mut self)

increment_le_inplace() treats the nonce as an unsigned little-endian number and increments it.

WARNING: this method does not check for arithmetic overflow. It is the callers responsibility to ensure that any given nonce value is only used once. If the caller does not do that the cryptographic primitives in rust_sodium will not uphold any security guarantees.