Module rust_sodium::utils

Libsodium utility functions

Functions

increment_le
increment_le() treats x as an unsigned little-endian number and increments it.

memcmp
memcmp() returns true if x[0], x[1], ..., x[len-1] are the same as y[0], y[1], ..., y[len-1]. Otherwise it returns false.

memzero
memzero() tries to effectively zero out the data in x even if optimizations are being applied to the code.

pad
Tries to add padding to a sequence of bytes. If the block size is zero, or the padded buffer's length could overflow usize, this function returns Err. Otherwise, it returns Ok wrapping the padded byte array.

unpad
Attempts to remove padding from a byte sequence created via pad(). If the padding is nonexistent, invalid, or the block size does not match the blocksize argument of pad(), this returns Err.